Refactor prevent conflict in export module
